Start creating a worker service fixture
